reorganize input.css location
